Fixing a problem caused by not cleansing user input. Any leading and/or trailing spaces must not be considered as part of the file name, otherwice a missing file to pop up shows up.
Improved user feedback on input errors due to non-existing model file,
or incorrect destination directory. Added a visual warning when the
chosen destination does not exist yet. A combination of visual clues
is used: a) errorneous input changes the text color to red; 
b) generation buttons are disabled when model file is not accessible, 
or if the target directory is invalid; c) if the target directory path
does not exist, the target dir label is decorated with info-level icon.