Use a simpler string as the default type name in TypeSupport.
The old way was to use the IDL IfR ID for the generated TypeSupport interface.
This resulted in strings that look like "IDL:Scope/NameTypeSupport:1.0".
Since this DDS spec allows the implementation to pick any name, this commit
changes it to the simpler C++ scoped name, in this example Scope::Name.