Send acknack when reader is going away
The following sequence was observed in the StaticDiscovery test:

Writer             Reader
Send last sample
...
                   Process last sample
                   ...
                   Terminate
...
Wait for acks
...
Timeout

The goal is to avoid the timeout.  There are three options:
1. Decrease the WaitForAck timeout.
2. Add a delay to the reader so that it will receive a heartbeat from
the writer and send an acknack.
3. Have the reader send an acknack before it goes away.  If the writer
receives it then it will not timeout.

This commit implements option 3.

The StaticDiscoveryTest was modified to set the history depth when
available.