Translate log file names for remote target
PerlDDS attempts to unlink log files before a test and scan log files
for errors after the test.  Assume the test is running on a remote
target whose file system has been mounted in the local system.  The log
files will appear on the local file system but must be translated to the
mount point.