Add -DCPSDefaultAddress
Safety profile builds must be configured with an address for discovery
and transport.  One way to do this is to specify a local_address in the
configuration file.  To avoid (1) manually editing every test
configuration file and (2) parsing every configuration file to determine
if and how the local_address should appear, we introduced a
configuration variable for the common section (and command-line option)
called DCPSDefaultAddress.  The default address overrides any that would
be determined automatically but does not override a local_address
appearing in a file.

This commit introduces the following changes:
* Adds default_address to the Service_Participant.
* Causes Sedp to use the default address when computing locators for Spdp.
* Causes Rtps to use the default address when computing locators for Sedp.
* The test framework will set the default address when
  DOC_TEST_*_IP_ADDRESS is defined.