Support temporary files in test framework
Some tests, e.g., DCPS/LivelinessTest, use temporary files to synchronize the
various processes.  Most of these tests assume that the files are created
in current working directory.  For split testing, the temporary files
must appear in a directory shared by both the host and the target.
Tests will use the add_temporary_file to add per-process temporary files
to the framework.  The framework will unlink the file when added and
when the test framework finishes.