Using reference counting to manage TransportClient::PendingAssoc objects.
The manual managnement for the lifetime of TransportClient::PendingAssoc 
objects fails occasionally because there’s no guarantee a PendingAssoc
object is no longer referenced by another thread when it is about to be 
deleted.

TransportClient::initiate_connect_i() had some logic to detect whether 
a PendingAssoc object was deleted or not before it returned to the 
caller (TransportClient::PendingAssoc::initiate_connect). However, it 
made the return value “false” from TransportClient::initiate_connect_i() 
ambiguous because it could mean connection failure or successfully 
connected but the PendingAssoc object is deleted already. This made it 
hard for the caller to avoid accessing the deleted object.