Dissector: Basic Payload Dissection in WS2
This is the basic framework for payload dissection in wireshark2.
Involves adding a hf vector to the Sample_Manager Singleton, adding
fields as they encountered in Sample_Dissector, and merging them with
the constant DCPS fields and passing the result to wireshark when DCPS
is registered. Sample Fields are left with a int that is the wireshark
field id (hf_).

During packet dissect, Sample_Fields are passed the wireshark field id
and using that they can add data corresponding to their field. Payload
type is now set on opendds.sample.payload and payload contents are added
in a tree under that.

Work left as of writing this is: to fix/complete registration of non
string fields, involve composite types (seq, union, etc), clean up
Sample_Manager usage, check/stop possible leak of dynamically generated
field names and get working with tests other than Messenger.