Dissector: Handle invalid ITLs, mark missing ITLs
Handle Segfaults in sample dissectors from invalid ITLs.
This wouldn't detect if an ITL is valid, just check if a pointer
is definitely invalid during the payload dissection, inform the user
via the GUI that the error is most likely caused by an invalid ITL file
and exit as gracefully as possible (exit()).

Ideally it would just mark the packets as malformed and recover, but
that code is less than ideal (using set long jump from the SIGSEGV handler).

Also in this commit is using Wireshark's expert info to inform the user in
the GUI that it couldn't dissect a sample payload for other reasons:
No Topic to Track or No Dissector could be found for the Type.