Mark struct as Primary DCPS Data Type in ITL file
IDL compiler will add a note "is_dcps_data_type" to ITL types that are a
struct delcared as the primary datatype with the DCPS_DATA_TYPE pragma.
This is intended to help in the generation of a proper namespace for the
Wireshark 2 dissector as the primary data types will be the roots of the
sample payload namespaces.