Dissector: Namespace Resolution and Union Bug
Added Sample_Base as a Parent Class to Sample_Field and
Sample_Dissector. Sample_Base holds wireshark namespace and field for
every context that it's child class is used in.

Tweaked Array and Sequence Labels and Namespace. For now using
"_e" (for element) to access an element of the sequence or array.

Removed namespace debug messages.

Fixed bug in Union length calculation that caused String Key Test to
hang on dissection.