Fixes for tests/DCPS/Reconnect
This test hides the program output.
Huangming had changed it in 373f41cff7bb3bf3b3f1f93249e9a1560dc7cd9a
from collecting individual program output in a single file to having one
for each program, except the InfoRepo. I added an InfoRepo log file and did
some cleanup/small improvements.

For sub_init_crash the publisher has always (or at least in the commit
before the WeakRcHandle PR) waited for subscribers until the test script
timedout, killed it and reported it as an error. The fact that no more
subscribers are spawned seems to be intentional, as it just tries to see
if the publisher will crash but it was also hidden before the PR.
Now it is killed and it shouldn't report an error unless
the publisher returns a non zero status.