MulticastReceivedStrategy: Connect ACE_Event_Handler refcount to RcObject
This is an attempt to correct a Access Violation on Windows when
the ACE Windows Implementation Reactor for the Multicast Transport tries
tries to access the Transport's MulticastReceivedStrategy as part of it's
cleanup. The problem is that it has already been deleted earlier in the
destructor of the transport. This enables ACE_Event_Handler's Reference
counting policy and connects it to MulticastReceiverStrategy's RcObject
count.

This is based off RcEventHandler, which might or might not be able to be
used instead.