Dissector: Move Field Info into Field_Context
Put the Wireshark field struct into Field_Context to be collected
after the entire tree has been evaluated. Refactored almost all the code
to account for this change, which consists of initializing the protocol
tree in two passes. Dissectors now need to be able to be deleted
if the first pass fails.