Use Builtin Entity Types for BIT Domain Entities
In otherwords use c2/c7 in the BIT reader/writers GUIDs for DCPS.
Also marks the BIT with a different enitity type, but this might
have to change. This work is not complete in this commit, as it seems
only the BITs and BIT Publisher Readers are getting the correct entity
types at this time.