Dissector: Removed RCH, Fixed Warnings
Using Reference Count Handles was not working out as well as I liked, so
I have removed them, but left the double tree pass initialization so
that something similar (along the lines of the Proxy pattern) might be
able to be done in the future maybe.

Also got rid of proto_tree_add_*_format() warnings by inserting "%s" in
the parameters.