Problem: RTPS does not work behind NATing firewalls
Network address translation (NAT) breaks RTPS in at least two ways.
Assume that two participants want to communicate via RTPS over the
public Internet and both of them are behind a firewall/router that
performs NAT.  First, multicast cannot be used for discovery because
the firewall will most likely not forward multicast traffic and even
if it did, multicast is not supported on the public Internet.  Second,
the locators advertised by SPDP and SEDP are only valid on the local
LAN due to NAT.

Solution: Relay RTPS messages through a well-known server.  The server
uses a security property at the participant level to create virtual
multicast groups.  Whenever a participant in one of the groups sends a
message, all of the other participants in the group will receive the
message.  See the README.rst for more details.