Problem: OpenDDS does not thrive in environments without multicast
We are moving away from non-standard discovery (InfoRepo) and
transports (TCP, UDP, multicast, etc.) and moving toward RTPS
discovery and transport.  RTPS relies on multicast for discovery.
However, many environments, like the cloud and the public interest do
not support multicast.

Solution: Write a multicast to unicast repeater for discovery.  The
repeater runs on each host with a participant and subscribes to the
discovery multicast group.  When it receives a message from the
participant, it forwards it via unicast to other hosts running the
repeater.  When the repeater receives a message via its unicast
socket, it repeats it on the multicast group so the local participants
can receive the message.  The repeater only needs to be used for
participant discovery since SEDP and data transport can use unicast.

The list of hosts to be contacted by the repeater can be static or
dynamic.  Currently, the repeater supports gathering a list of hosts
in a Google Cloud Platform (GCP) instance group.