Empty Commit by Fred Hornsey
I'm making a change to gitrelease.pl to use git shortlog to gather the
author info for the AUTHORS file. This wouldn't be a problems except one
of the names currently listed in AUTHORS isn't in git's history:
  Tim Bradley <bradley_t@ociweb.com>

To remedy this, I'm adding this empty commit with Tim listed as the
author so he'll always show up.