Fix Auth/Key Exchange "Race Condition"
In security, there is a three way handshake which makes up
authentication and precedes key exchange. The participant that is
predetermined to be the initiator will finish the handshake first and
will continue on to key exchange. If the speeds are matched between the
two participants this is not a problem. However if the non-initiator is
much slower, then as of writing it is possible that the volatile reader
that does key exchange will not be ready to decrypt messages. Because of
how durability works reliability doesn't help because key exchange
happened before the reader could see anything.

The solution used here is to resend keys on the first AckNack from the
key exchange reader, after we know the reader can decrypt them. They are
the same keys in new messages so it gets around the problem caused by
volatile durability.