Problem: Inconsistent topic only discovered with local reader/writer
A participant can only get inconconsistent topic when it has a local
reader or writer on that topic.  While usual, a participant should be
able to create a topic and then get inconsistent topic on it without
creating readers and writers.  This also lead to a race condition in
earlier versions of the InconsistentTopic test which made it fail
sporadically.

Solution: Track local and remote topics and link the local topic to
its impl through callbacks so that inconsistent topic can be reported
without a reader or writer.