Shutdown fails when Service Participant thread is interrupted
An interrupt delievered to the Service Participant thread causes the
run_reactor_event_loop function to return with an "interrupted system
call" error.  An attempt to shutdown will notify the now stopped
reactor.  Since the reactor isn't running,
DomainParticipantImpl::handle_exception is never called.  Since
handle_exception is never called, the condition variable is not
released and shutdown hangs.

The solution is to block signals in the Service Participant thread.