Reactor task in Service Participant has stale reference
A ReactorTask is an ACE::Task_Base which caches a references to its
Thread_Manager.  If the Service_Participant is shutdown and ACE is
restarted (ACE::fini, ACE::init), then this reference is stale and
shows up as an access violation in the Restart test.

Solution: Reset the reference when the ReactorTask is stop so that it
will use the correct Thread_Mananger when restarted.