ReactorInterceptor wait is broken
To support a straight enqueue while processing commands in the
ReactorInterceptor, it was necessary to release the queue lock before
each command was executed.  This broke the existing semantics of wait
because the following sequence was possible:

Non-reactor thread: execute_or_enqueue (command is placed on the
queue)
Reactor thread: command is take off the queue but not executed yet
Non-reactor thread: wait (returns immediately since the queue is
empty)
... wait returned without the command being executed ...

Solution: enqueue and execute_or_enqueue return a Command with a wait
method.  A thread can use this wait on a specify command to complete.
The ReactorInterceptor signals the Command to unblock any waiting thread.