Modified GuidGenerator and added sequence reset
GuidGenerator Code was modified to add a random generated number
to the counter value.  The counter value will increment from the
initial random generated number.

Added RTPSDiscovery configuration value
MaxSpdpSequenceMsgResetChecks.  This value is the number of sequence
numbers that must be received that are less than the initial last
sequence number value before a discovered participant is removed.

Added Spdp code changes to check for a sequence number reset. A reset
occurs when a defined number of sequence numbers are received that are
less than than an initial last sequence value.
On a reset the discovered participant will be removed and rediscovered.