SPDP Sequence Check
This code checks for a possible participant reset.  If a participant
already exists and the code begins receiving sequence numbers less than
the last sequence number.  The participant will be removed if we recieve
a configurable number of sequences less than the last sequence.  The
participant then will be re-discovered.