Added an optional argument, TARGET_LINK_KEYWORD, in the cmake macro OPENDDS_TARGET_SOURCES that will allow the user the option to specify the scope as PUBLIC_LINK|PRIVATE_LINK|INTERFACE_LINK for the libraries that are linked against in the macro. The keyword is made up of the scope, appended with _LINK, to distinguish the scope that is used in linking from the scope that maybe specified for the sources.
If the user does not specify scope by use of the keyword, the plain signature for the target will be used when target_link_libraries(...) is called in the macro, and because of cmake policy CMP0023, all uses of target_link_libraries for the target must use the plain signature, else a cmake error will be encountered. Similarly, if TARGET_LINK_KEYWORD is specified, all other uses of target_link_libraries for that target must make use of the keyword signature.