Relay addresses not updated when client changes servers
Previously, the relay addresses were bundled with the reader and
writer entries.  Suppose there are two relays A and B and a client
connects to A, switches to B, and then switches back to A before the
participant lease duration.  This causes A to write entries for the
participant and then B will write entries for the participant.
However, the client is associated with A but the topic says it should
be associated with B.

To solve this problem, remove the relay addresses from the reader and
writer entries and create a new topic for participant to relay
pairings (Responsible Relay).  The relays maintain the Responsible
Relay topic based on the traffic that they are observing.  The Reader
and Writer entries are just for matching.

For this change, we had to remove the local/remote distinction in the
QosIndex.  This means that every relay must maintain a complete
association table.