Spdp uses destructed lock on shutdown (#1354)
* Spdp uses destructed lock on shutdown

The SPDP transports sends unregister/dispose messages as part of
shutdown.  If it is using the RtpsRelay, then it must acquire the lock
for the address.  However, the lock has already been destroyed.

The solution was to move all of the configuration for RtpsDiscovery to
a separate configuration object that will not be destroyed until the
Spdp object is destroyed.