RTPS discovery and transport cannot handle network changes
The locators determined by RTPS discovery and used by the RTPS
transport are set at start time.  This prevents OpenDDS from be used
in situations where the network configuration is expected to change
like mobile phones, tablets, etc.

To solve this problem, code was added to detect network changes and
update the locators announced by SPDP and SEDP.  Support was also
added for checking for changed locators and updating the RTPS
transport to use the new locators.  ICE was also updated so that the
published AgentInfo is updated when the network configuration changes.

Only Linux is supported at this time.

Diagnostic logging for these features can be turned on with
DCPSDebugLevel > 4.