ICE fails to converge promptly
Foundations used by connectivity checks that fail were being removed
from the set of active foundations.  However, this was not being
propagated to other checklists that were frozen on that foundation.
This change unfreezes checklists after a foundation is removed from
the set of active foundations.

This change revealed a deadlock which was resolved by queuing STUN
sends in SPDP.