Use reference counting for ICE
The time-based behavior of Checklists, ServerReflexiveTasks, and
ChangePasswordTasks is governed by the AgentImpl.  There was no way to
remove a Task from the queue of Tasks waiting on a timer.
Consequently, manual memory management was used to 1) prevent a Task
from entering the queue twice and 2) to destroy the Task as a
scheduled event.  The problem is the ServerReflexiveTask and
ChangePasswordTask have the same lifetime as their parent
object (EndpointManager) but the destruction was solely under the
control of the ServerReflexiveTask.  Thus, a destroyed
ChangePasswordTask may be in the queue of Tasks waiting on a timer.

Solution: Change the queue of Tasks to hold weak references.