Try to join IPV4 and IPV6 multicast groups when no NCM
Problem: The logic to join multicast groups on platforms that don't
have a network config monitor (NCM) creates a default NIC and populates it
with an address to be compatible with the NCM logic.  If IPV6 is
enabled, then the default address will be an IPV6 address meaning that
the IPV4 multicast groups will never be joined.

Solution: Always add an IPV4 address to the default NIC and add an
IPV6 address if appropriate.