Remove OfferedDeadlineWatchdog
The watchdog uses timer_ids and therefore must wait() when scheduling
timers.  This effectively introduces a lock which leads to deadlock
when the Deadline QoS is used.  The deadly embrace occurs between the
writing thread which has the DataWriterImpl lock and needs the reactor
lock and the Reactor thread servicing the watchdog which has the
reactor thread and needs the DataWriterImpl lock.

Solution: Rewrite the OfferedDeadlineWatchdog using SporadicTask.

In general, wait()ing on Command objects in the ReactorInterceptor is
unsafe and the wait functionality should be removed to reduce the
chance of similar deadlocks in the future.