Only send auth replies for auth requests
One participant (the initiator) has discovered another
participant (the replier) and sent an auth request.  Due to SPDP
reset, it removes the participant but immediately rediscovers it and
sends another auth request.  If the replier receives the first auth
request, it will continue to respond to it but authentication will
fail as the initiator has changed the Diffie-Hellman parameters.

Solution:

1. The replier will not periodically resend replies.  It will only
   send auth replies in response to an auth request.  It will send an
   auth reply for every request.
2. If the replier detects that the auth request has changed, it will
   recompute a new auth reply.