Unlock for ParticipantLocationTopic unregister/dispose
This was motivated by a deadlock.  The first thread was adding a
participant on the basis of a successful security handshake.  This
thread unlocks the SPDP lock so that it can call into the
ParticipantLocationTopic which acquires the sample lock and then
reacquires the SPDP lock.  The second thread was sending a periodic
announcement and, as a by-product, cleaning up expired participants.
This thread acquired the SPDP lock before calling into DiscoveryBase
to remove the participant.  The code in discovery base called into the
ParticipantLocationTopic with the SPDP lock.  The second thread then
attempts to acquire the sample lock.

Solution: Unlock the SPDP lock before interacting with the
ParticipantLocationTopic reader when removing a discovered participant.