Associations are removed after expire-associate
Suppose that a Participant A is suspended long enough that it expires
for Participant B.  Participant A is then resumed.  In RTPS, this will
most likely cause a number of timers to fire include the one that will
cause Participant A to 1) expire Participant B and 2) send SPDP.  As
part of expiring Participant B, associations between writers in
Participant B and readers in Participant A will be scheduled for
removal (in Participant A).  Sending SPDP will restart the discovery
process.  This may cause associations between the readers and writers
to complete for the timer fires that causes the association to be
removed.  Finally, that timer will fire and remove the association
that was just added.

Solution: When adding an association to a reader, cancel any
RemoveAssociationSweeper that will remove the writer.