Readers/writers in same secure participant do not communicate
When using RTPS, readers and writers in the same secure participant
exchange data through RTPS.  For this to work, the
RtpsUdpReceiveStrategy must have the participant's crypto handle.
Otherwise, the message will not be decrypted and warnings will be
generated to this effect.

Solution:  Add the necessary crypto handles when the readers and
writers in the same local participant.