Participants cannot reauthenticate
Suppose participant A has a lease time of 100 seconds and participant
B has a lease time of 50 seconds.  Participant B is suspended for 50+
seconds so that it expires in Participant A.  Participant B is then
resumed.  This will cause Participant A to rediscover Participant
B.  (Participant B does not need to rediscover Participant A because
it has not expired.)  Participant A must reauthenticate with
Participant B.

If Participant A is the requester, then Participant B must restart
authentication when it receives the handshake request.  If Participant
B is the requester, then Participant A must send an auth request and
Participant B must restart the authentication processes.

Solution: Add support for both reauthentication scenarios.