Allow Built-in Security Plugin to be Reloaded
Fix for static secure restart test.

This makes BuiltInPluginLoader check if it needs to act if called again
instead of just acting once. Then I added that a call to the loader to
the Restart test.

An Issue:
The last commit appears to have a chance to segfault the restart test on
the ending of a Application object. I have no idea how it's doing this.

One Question:
If we're static, is it wise to try to reload the built-in security
plugin automatically from SecurityRegistry::load_security_plugin_lib?