Fix Default Allowed Data Representation
Set the allowed data representation of types without a annotation to all
of them, or all true. I might have gotten mixed up with the fact that in
DataRepresentationAnnotation::node_value_exists, when there is an
annotation, the starting value has to be all false and then the
annotations add values to it.