RTPS Discovery does not degrade gracefully with load
SEDP messages are queued and processed on a different thread than the
SEDP transport.  This queue has blocking behavior and causes a
deadlock when it fills up.

Solution:  The solution is to create back pressure at the point of
input processing.  This means that the transport thread(s) that process
the input should be doing the work of processing the messages.  This
commit shifts all discovery processing to the SEDP transport thread
and eliminates the SEDP worker thread and SPDP transport thread.