Incorrect flags for Participant Message Data Writer
Methods using the Participant Message Data Writer (participant
liveliness) were using the wrong flag to see if it was enabled.  If
the writers were not enabled, the incorrect flag was not set which
cause the writer to not be associated with a remote reader.
Consequently, the reader would send acknacks that were never answered.

Solution:  Use the correct flag to check if the Participant Message
Data Writer is enabled.  Update the not enabled writers setting to more
accurately match the desired use case of a participant that does not
write about publications, subscriptions, or liveliness.