User cannot dynamically toggle RtpsRelay and ICE
The decision to use the RtpsRelay (or to only use the RtpsRelay) or
ICE can only be made at start time.  Users of OpenDDS want the ability
to dynamically turn off the RtpsRelay and ICE for testing purposes.

Solution: Make the corresponding configuration dynamic.  The
configuration for the RtspRelay is straightforward as it can be picked
up on the next send.  The configuration of ICE is more difficult since
ICE needs to be restarted for participants that have already been
discovered.

The code attempts to reset flags in the ParticpantLocationTopic.  The
ParticipantLocationTopic in a remote participant will be updated after
the corresponding entries time out.  To a certain extent, these
controls influence the ways in which other participants can discover a
given participant.  For example, switching to RtpsRelay only means
that a remote participant will eventually say that the given
participant can only be reached via the RtpsRelay.  The local
participant may still be able to discover the remote locally.