NAT bindings for RtpsRelay may timeout without return data
A client was observed sending SEDP to the relay and the relay was
observed sending SEDP to the client.  The client's messages were
received by the relay but the relay's messages were not received by
the client.  Simultaneously, there was no issue with the SPDP and STUN
messages to the SEDP port.  The key difference between the SEDP port
and SPDP/STUN is that the latter has a periodic return of data.  Thus,
the motivation for this change is to provide a mechanism for the
periodic return of data from the relay to a client.

Solution
--------

Each port on the relay acts as a STUN server.  Clients of the relay
periodically send binding requests to refresh any NAT bindings.
Client send binding indications between requests.  The timing of the
requests is configured through the ICE setting for server reflexive
addresses.

The STUN messages to the RtpsRelay were extended with the GUID prefix
of the client.  This was necessary to setup and maintain the routing
tables on the relay.

A new builtin topic (Connection Record) was created to monitor the
connection with the relay.

Future Work
-----------

The Connection Record topic should replace the Participant Location
Topic.

When the RtpsRelay and ICE are used simultaneously, the server
reflexive task in ICE should be disabled and the results of the relay
STUN task used instead.

The next possible step in this design is to implement STUN via RTPS.
One would most likely take the route of STUN readers and writers and
secure versions of the same.  One challenge is that the ports in the
relay are not associated with a participant.  One option is for the
relay to impersonate the application participant for the purpose of
reading and writing STUN.