Observed data cannot be logged
Problem
-------

One of the primary motivations for the Observer interface is the
ability to log samples as they are written, received, and accessed.
Currently, users of the Observer interface receive a generic pointer
to sample data which may be a raw sample or something else.
There is no capability that allows a user to take this generic
pointer and log the sample contained therein.

Solution
--------

1. Standardize on the approach that the pointer in the Sample is a
pointer to a raw sample.

2. Introduce the ValueWriter interface and necessary support in the
IDL compiler to allow generic processing of the raw sample.

3. Add plumbing between generic and typed DataReader and DataWriter to
allow a user to dispatch a ValueWriter on the raw sample.

4. Provide an incomplete JsonValueWriter to demonstrate how types
processed by the IDL compiler could be serialized to JSON.

Notes
-----

* It may be necessary/beneficial to augment the ValueWriter interface
  with type information via TypeObjects.

* ValueWriter suggests a symmetric interface called ValueReader that
  takes an opaque representation and populates a raw sample, e.g.,
  deserialization.