Publications and Subscriptions BITs don't work in RTPS rediscovery
Problem
-------

Consider the following scenario:

* Participant A and B have discovered each other.
* A network partition or other issue that causes participants to
  timeout of discovery occurs.
* The issue resolves and A and B discovery each other again.

Insertion into the Publication and Subscription BITs will fail when A
and B discover each other again.  The core of the problem is that SEDP
will assign a new BIT key to the Publication/Subscription while the
DataReader is expecting the same key since the instance may have not
been removed from the BIT.

Solution
--------

Adopt the resolution of issue 153 for the DDS 1.5 Spec that says that
BIT keys are 16 byte arrays.  This allows a BIT key to store an RTPS
GUID.