BIT topics are spec-defined memory leak
Problem
-------

The DDS Specification says that the members of the
ReaderDataLifecycleQosPolicy are set to infinite which means a
participant will never purge the data stored in the builtin topics.  A
long-running participant will eventually run out of memory.

It is possible to manually change the qos for all of the builtin
participant data readers.  However, the addition of a new builtin
topic would render such code useless and the new builtin topic would
create another memory leak.

Solution
--------

Provide a mechanism to configure the ReaderDataLifecycleQosPolicy for
the builtin topic data readers.