Deque is memory inefficient
Problem
-------

Currently, a typical DDS application might involve 50+
ReactorInterceptors.  Each ReactorInterceptor includes a deque.  A
deque is built from internal nodes.  The internal nodes typically have
a minimum size which is the lesser of the size of the value type and
some constant.  This constant can be much bigger than the size of the
value type.  Consequently, a deque that contains a only a few entries,
such as one for a ReactorInterceptor, is highly memory inefficient.
If an empty deque is 1K, then a typical DDS application will have 50K+
of memory that is not actually being used.

Solution
--------

Change ReactorInterceptor and Job queue to use vectors that are
swapped with empty vectors as they are being processed.