Secure discovery is slow
Problem 1
---------

The default configuration delays an SPDP resend for 3 seconds.
However, a participant A receiving an initial SPDP announcement from B
will attempt authentication with B.  However, B will ignore the
authentication messages because it has not discovered A.

Solution 1
----------

Send an SPDP announcement when attempting authentication with a newly
discovered participant.

Problem 2
---------

`match_authenticated` is called before the FINAL in the auth handshake
is sent.  This is problematic because `match_authenticated` is slow,
e.g., takes 1 second to complete.  The other participant is idle while
waiting for the FINAL and will call `match_authenticated` when it gets
the FINAL.

Solution 2
----------

Send the FINAL before calling `match_authenticated`.  While it doesn't
address the reason that `match_authenticated` is so slow, it helps
because both participants can be in `match_authenticated` at the same
time.

Problem 3
---------

`match_authenticated` associates all of the builtin readers and
writers.  At more-or-less the same time, it is sending crypto tokens
for the builtin endpoints.  This creates a race where the heartbeats
and acknacks for reliability can get dropped because the crypto tokens
are not in place to decrypt them.  This means discovery has to wait
one or two rounds of reliability which is 1 or 2 seconds by default.

Solution 3
----------

Defer associating the builtin endpoints until the crypto tokens have
been received for the remote.  This suggests that a similar race
condition may exist for user endpoints.