RTPS heartbeats are not scalable
Problem
-------

The current algorithm for sending heartbeats makes a copy of the
writers data structure and then has a double for-loop iterating over
all of the writers and their readers.  If the writers data structure
is large, then copying and brute-force iteration is prohibitively
slow.  Furthermore, the participant volatile secure (PVS) writer for
security compounds the problem since each new participant advances the
sequence number but all existing participants do not care.

Solution
--------

For PVS, maintain a per-reader sequence number that is the expected
cumulative ack - 1 for that reader.

Maintain a set of readers for each writer that requires a heartbeat.